Updated Debug logging

+++ src/ElementBuffer.h
@@ -16,7 +16,7 @@ public:
     ElementBuffer& operator=(ElementBuffer&&) noexcept;
     ~ElementBuffer();
     
-    void CreateEBO(unsigned int* data, unsigned int& dataSize, GLenum mode);
+    void CreateEBO(GLuint* data, size_t& dataSize, GLenum mode);
     void Bind() const;
     void Unbind() const;
 };
