+++ src/VertexBuffer.h
@@ -16,7 +16,7 @@ public:
     VertexBuffer& operator=(VertexBuffer&& other) noexcept;
     ~VertexBuffer();
     
-    void CreateVBO(float* data, unsigned int& dataSize, GLenum mode);
+    void CreateVBO(float* data, size_t& dataSize, GLenum mode);
     void Bind() const;
     void Unbind() const;
 };
